Rename PeerList to SyncStatus
